changed: handling of file extension + code cleanup
removed functions in CUtil where input value was a CStdString and return value a pointer to the internal buffer in CStdString.. very easy to mess up.
moved getextension code into curl, as normally you need to parse the url to get the extension since periods in folders should not be considered extensions.


git-svn-id: http://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@5479 568bbfeb-2a22-0410-94d2-cc84cf5bfa90